requirements.txt
pandas
openpyxl
python-jobspy
pyyaml
sentence-transformers
torch
